requirements.txt
Pillow>=10.0.0  # Python Imaging Library fork, needed for image processing
